Separati in file http server, app express, e socket.io
